# backend/src/main/java/pl/twinaxeshop/dao/response/JwtAuthenticationResponse.java
package pl.twinaxeshop.dao.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class JwtAuthenticationResponse {
    private String token;
}